moved ComputeNumberMatched to common config package
